convert int32 in float32 because pandas converts int32 in int64 which causes memory problems

+++ src/core/columns.py
@@ -42,7 +42,7 @@ class IntCol(Column):
     '''
     def __init__(self, label = None, default = 0):
         super(IntCol, self).__init__(label, default)
-        self._dtype = np.int32
+        self._dtype = np.float32
         
 class EnumCol(IntCol):
     '''
